finished

+++ doubly_linked_list/doubly_linked_list.py
@@ -49,8 +49,7 @@ class DoublyLinkedList:
     def remove_from_head(self):
         if self.head:
             value = self.head.value
-            self.head.delete()
-            self.length -= 1
+            self.delete(self.head)
             return value
     
     """
@@ -73,19 +72,17 @@ class DoublyLinkedList:
     Returns the value of the removed Node.
     """
     def remove_from_tail(self):
-         if self.tail:
+        if self.tail:
             value = self.tail.value
-            self.tail.delete()
-            self.length -= 1
+            self.delete(self.tail)
             return value
-
     """
     Removes the input node from its current spot in the 
     List and inserts it as the new head node of the List.
     """
     def move_to_front(self, node):
         node_val = node.value
-        node.delete()
+        self.delete(node)
         self.add_to_head(node_val)
         
     """
@@ -94,7 +91,7 @@ class DoublyLinkedList:
     """
     def move_to_end(self, node):
         node_val = node.value
-        node.delete()
+        self.delete(node)
         self.add_to_tail(node_val)
 
     """
@@ -102,7 +99,18 @@ class DoublyLinkedList:
     order of the other elements of the List.
     """
     def delete(self, node):
-        node.delete()
+        if not self.head and not self.tail:
+            return
+        self.length -=1
+        if self.head is self.tail:
+            self.head = None
+            self.tail = None
+        elif self.head == node:
+            self.head = node.next
+        elif self.tail == node:
+            self.tail = node.prev
+        else:
+            node.delete()
 
     """
     Finds and returns the maximum value of all the nodes
